Refactor Garmin scripts: Rename 'runs' to 'cardio' and harden path handling for cron usage
