+++ daily_garmin_health.py
@@ -11,7 +11,8 @@ import platform
 from dotenv import load_dotenv
 
 # 1. Load configuration immediately
-load_dotenv()
+SCRIPT_DIR = os.path.dirname(os.path.abspath(__file__))
+load_dotenv(os.path.join(SCRIPT_DIR, ".env"))
 
 # 2. Get the settings (with defaults for safety)
 # On Raspberry Pi/Linux: Set CHECK_MOUNT_STATUS=True in .env to enable mount verification
@@ -37,7 +38,6 @@ elif check_mount and is_windows:
 # ... rest of your code ...
 
 # --- CONFIGURATION VIA ENVIRONMENT ---
-load_dotenv()
 SAVE_PATH = os.getenv("SAVE_PATH")
 
 if SAVE_PATH:
@@ -46,7 +46,7 @@ else:
     print("WARNING: SAVE_PATH not set in .env. Using current folder.")
     CSV_FILE = "garmin_stats.csv"
 
-TOKEN_DIR = ".garth"
+TOKEN_DIR = os.path.join(SCRIPT_DIR, ".garth")
 # -------------------------------------
 
 def get_safe(data, *keys):
